Split ball and followcircle into default/legacy files
