+++ osu.Game.Rulesets.Osu/Objects/Drawables/DrawableSliderBall.cs
@@ -7,24 +7,21 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using osu.Framework.Allocation;
-using osu.Framework.Bindables;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
-using osu.Framework.Graphics.Shapes;
 using osu.Framework.Input;
 using osu.Framework.Input.Events;
 using osu.Game.Graphics;
 using osu.Game.Rulesets.Objects.Drawables;
 using osu.Game.Rulesets.Objects.Types;
-using osu.Game.Rulesets.Osu.Objects;
-using osu.Game.Rulesets.Osu.Objects.Drawables;
+using osu.Game.Rulesets.Osu.Skinning.Default;
 using osu.Game.Skinning;
 using osuTK;
 using osuTK.Graphics;
 
-namespace osu.Game.Rulesets.Osu.Skinning.Default
+namespace osu.Game.Rulesets.Osu.Objects.Drawables
 {
-    public class SliderBall : CircularContainer, ISliderProgress, IRequireHighFrequencyMousePosition, IHasAccentColour
+    public class DrawableSliderBall : CircularContainer, ISliderProgress, IRequireHighFrequencyMousePosition, IHasAccentColour
     {
         public Func<OsuAction?> GetInitialHitAction;
 
@@ -34,14 +31,15 @@ namespace osu.Game.Rulesets.Osu.Skinning.Default
             set => ball.Colour = value;
         }
 
-        private readonly Drawable followCircle;
-        private readonly Drawable fullSizeFollowCircle;
-        private readonly DrawableSlider drawableSlider;
-        private readonly Drawable ball;
+        private Drawable followCircle;
+        private Drawable followCircleReceptor;
+        private DrawableSlider drawableSlider;
+        private Drawable ball;
 
-        public SliderBall(DrawableSlider drawableSlider)
+        [BackgroundDependencyLoader]
+        private void load(DrawableHitObject drawableSlider)
         {
-            this.drawableSlider = drawableSlider;
+            this.drawableSlider = (DrawableSlider)drawableSlider;
 
             Origin = Anchor.Centre;
 
@@ -49,15 +47,14 @@ namespace osu.Game.Rulesets.Osu.Skinning.Default
 
             Children = new[]
             {
-                followCircle = new FollowCircleContainer
+                followCircle = new SkinnableDrawable(new OsuSkinComponent(OsuSkinComponents.SliderFollowCircle), _ => new DefaultFollowCircle())
                 {
                     Origin = Anchor.Centre,
                     Anchor = Anchor.Centre,
                     RelativeSizeAxes = Axes.Both,
                     Alpha = 0,
-                    Child = new SkinnableDrawable(new OsuSkinComponent(OsuSkinComponents.SliderFollowCircle), _ => new DefaultFollowCircle()),
                 },
-                fullSizeFollowCircle = new CircularContainer
+                followCircleReceptor = new CircularContainer
                 {
                     Origin = Anchor.Centre,
                     Anchor = Anchor.Centre,
@@ -106,7 +103,7 @@ namespace osu.Game.Rulesets.Osu.Skinning.Default
 
                 tracking = value;
 
-                fullSizeFollowCircle.Scale = new Vector2(tracking ? 2.4f : 1f);
+                followCircleReceptor.Scale = new Vector2(tracking ? 2.4f : 1f);
 
                 followCircle.ScaleTo(tracking ? 2.4f : 1f, 300, Easing.OutQuint);
                 followCircle.FadeTo(tracking ? 1f : 0, 300, Easing.OutQuint);
@@ -167,7 +164,7 @@ namespace osu.Game.Rulesets.Osu.Skinning.Default
                 // in valid time range
                 Time.Current >= drawableSlider.HitObject.StartTime && Time.Current < drawableSlider.HitObject.EndTime &&
                 // in valid position range
-                lastScreenSpaceMousePosition.HasValue && fullSizeFollowCircle.ReceivePositionalInputAt(lastScreenSpaceMousePosition.Value) &&
+                lastScreenSpaceMousePosition.HasValue && followCircleReceptor.ReceivePositionalInputAt(lastScreenSpaceMousePosition.Value) &&
                 // valid action
                 (actions?.Any(isValidTrackingAction) ?? false);
 
@@ -204,75 +201,6 @@ namespace osu.Game.Rulesets.Osu.Skinning.Default
 
             ball.Rotation = -90 + (float)(-Math.Atan2(diff.X, diff.Y) * 180 / Math.PI);
             lastPosition = Position;
-        }
-
-        private class FollowCircleContainer : CircularContainer
-        {
-            public override bool HandlePositionalInput => true;
-        }
-
-        public class DefaultFollowCircle : CompositeDrawable
-        {
-            public DefaultFollowCircle()
-            {
-                RelativeSizeAxes = Axes.Both;
-
-                InternalChild = new CircularContainer
-                {
-                    RelativeSizeAxes = Axes.Both,
-                    Masking = true,
-                    BorderThickness = 5,
-                    BorderColour = Color4.Orange,
-                    Blending = BlendingParameters.Additive,
-                    Child = new Box
-                    {
-                        Colour = Color4.Orange,
-                        RelativeSizeAxes = Axes.Both,
-                        Alpha = 0.2f,
-                    }
-                };
-            }
-        }
-
-        public class DefaultSliderBall : CompositeDrawable
-        {
-            private Box box;
-
-            [BackgroundDependencyLoader]
-            private void load(DrawableHitObject drawableObject, ISkinSource skin)
-            {
-                var slider = (DrawableSlider)drawableObject;
-
-                RelativeSizeAxes = Axes.Both;
-
-                float radius = skin.GetConfig<OsuSkinConfiguration, float>(OsuSkinConfiguration.SliderPathRadius)?.Value ?? OsuHitObject.OBJECT_RADIUS;
-
-                InternalChild = new CircularContainer
-                {
-                    Masking = true,
-                    RelativeSizeAxes = Axes.Both,
-                    Scale = new Vector2(radius / OsuHitObject.OBJECT_RADIUS),
-                    Anchor = Anchor.Centre,
-                    Origin = Anchor.Centre,
-                    Blending = BlendingParameters.Additive,
-                    BorderThickness = 10,
-                    BorderColour = Color4.White,
-                    Alpha = 1,
-                    Child = box = new Box
-                    {
-                        Blending = BlendingParameters.Additive,
-                        RelativeSizeAxes = Axes.Both,
-                        Colour = Color4.White,
-                        AlwaysPresent = true,
-                        Alpha = 0
-                    }
-                };
-
-                slider.Tracking.BindValueChanged(trackingChanged, true);
-            }
-
-            private void trackingChanged(ValueChangedEvent<bool> tracking) =>
-                box.FadeTo(tracking.NewValue ? 0.3f : 0.05f, 200, Easing.OutQuint);
         }
     }
 }
